refactor: move file loading logic from sessions to dedicated io module
